# custom_addons/pitch_api/static/doc/openapi.yaml
openapi: 3.0.0
info:
  title: Pitch API
  version: "1.0.0"
  description: |
    OpenAPI 3.0 specification for the `pitch_api` module. Covers Authentication
    endpoints (login, refresh, logout, introspect, revoke refresh, devices, me)
    and Appointment endpoints (single, list, paginated, today, app_screen_logs).
servers:
  - url: http://{host}
    description: Replace {host} with your Odoo host
    variables:
      host:
        default: localhost
        description: Odoo host
paths:
...
  /api/users:
    get:
      tags: [users]
      summary: List all users (admin-only)
      description: List all users with key profile fields. Groups information is available via separate endpoints. Requesting user must be a Pitch API Admin.
      security:
        - bearerAuth: []
      parameters:
        - name: active
          in: query
          required: false
          description: |
            Filter by user status:
            - active=true: only active users
            - active=false: only inactive users
            - If not provided: returns ALL users (both active and inactive)
          schema:
            type: string
            enum: [true, false]
          example: "true"
      responses:
        '200':
          description: A list of users.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{user_id}:
    get:
      tags: [users]
      summary: Get a user by ID (admin-only)
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/lookup:
    get:
      tags: [users]
      summary: Lookup a user by ID or login (admin-only)
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          schema:
            type: integer
        - name: login
          in: query
          schema:
            type: string
      responses:
        '200':
          description: User object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/exists:
    get:
      tags: [users]
      summary: Check if a user exists (admin-only)
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          schema:
            type: integer
        - name: login
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Existence check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{user_id}/groups:
    get:
      tags: [users]
      summary: Get groups for a single user (admin-only)
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of groups for the specified user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGroupsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/groups:
    get:
      tags: [users]
      summary: Get groups for multiple users (admin-only)
      security:
        - bearerAuth: []
      parameters:
        - name: user_ids
          in: query
          required: true
          description: Comma-separated list of user IDs
          schema:
            type: string
      responses:
        '200':
          description: A list of users with their respective groups.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiUserGroupsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        error_description:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        login:
          type: string
        active:
          type: boolean
        tz:
          type: string
        company_id:
          type: integer
          nullable: true
        company_name:
          type: string
          nullable: true
        login_date:
          type: string
          format: date-time
          nullable: true
        is_pitch_admin:
          type: boolean
        improveit_user_id:
          type: string
          nullable: true
    UserListResponse:
      type: object
      properties:
        admin_user_id:
          type: integer
        count:
          type: integer
          description: Number of users returned
        active_filter:
          type: string
          description: "Applied active filter value. Shows 'all' when no active parameter provided, otherwise shows the actual parameter value (true/false)."
          example: "all"
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
    UserGroup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        xml_id:
          type: string
          nullable: true
        category_id:
          type: integer
          nullable: true
        category_name:
          type: string
          nullable: true
    UserGroupsResponse:
      type: object
      properties:
        user_id:
          type: integer
        user_login:
          type: string
        user_name:
          type: string
        groups_count:
          type: integer
        groups:
          type: array
          items:
            $ref: '#/components/schemas/UserGroup'
    MultiUserGroupsResponse:
      type: object
      properties:
        count:
          type: integer
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserGroupsResponse'
...